requirements.txt
markdown
djangorestframework
